refactor(#114): post 관련 클래스 통합
