+++ src/main/java/deepdivers/community/domain/post/controller/PostApiController.java
@@ -10,11 +10,13 @@ import deepdivers.community.domain.post.dto.request.ProjectPostRequest;
 import deepdivers.community.domain.post.dto.response.PostSaveResponse;
 import deepdivers.community.domain.like.service.LikeService;
 import deepdivers.community.domain.post.service.PostService;
+import deepdivers.community.domain.post.service.ProjectPostService;
 import deepdivers.community.global.security.Auth;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -28,6 +30,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class PostApiController implements PostApiControllerDocs {
 
 	private final PostService postService;
+	private final ProjectPostService projectPostService;
 
 	@Override
 	@PostMapping("/upload")
@@ -58,6 +61,34 @@ public class PostApiController implements PostApiControllerDocs {
 	) {
 		final NoContent response = postService.deletePost(postId, member);
 		return ResponseEntity.ok(response);
+	}
+
+	@PostMapping("/project/upload")
+	public ResponseEntity<API<Long>> createProject(
+		@Auth final Member member,
+		@Valid @RequestBody final ProjectPostRequest dto
+	) {
+		final API<Long> result = projectPostService.createProjectPost(member, dto);
+		return ResponseEntity.ok(result);
+	}
+
+	@PostMapping("/project/edit/{projectId}")
+	public ResponseEntity<API<Long>> editProject(
+		@Auth final Member member,
+		@Valid @RequestBody final ProjectPostRequest dto,
+		@PathVariable final Long projectId
+	) {
+		final API<Long> result = projectPostService.updateProjectPost(projectId, member, dto);
+		return ResponseEntity.ok(result);
+	}
+
+	@DeleteMapping("/project/remove/{projectId}")
+	public ResponseEntity<NoContent> removeProject(
+		@Auth final Member member,
+		@PathVariable final Long projectId
+	) {
+		final NoContent noContent = projectPostService.deletePost(projectId, member);
+		return ResponseEntity.ok(noContent);
 	}
 
 }
