+++ src/main/java/deepdivers/community/domain/post/dto/response/ProjectPostDetailResponse.java
@@ -1,5 +1,10 @@
 package deepdivers.community.domain.post.dto.response;
 
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
 public class ProjectPostDetailResponse extends ProjectPostPreviewResponse {
 
     private boolean isLikedMe;
