+++ src/main/java/deepdivers/community/domain/post/repository/ProjectPostQueryRepositoryImpl.java
@@ -4,14 +4,12 @@ import static deepdivers.community.domain.like.entity.QLike.like;
 import static deepdivers.community.domain.member.entity.QMember.member;
 import static deepdivers.community.domain.post.entity.QPost.post;
 
-import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import deepdivers.community.domain.category.entity.CategoryType;
 import deepdivers.community.domain.common.exception.NotFoundException;
 import deepdivers.community.domain.file.application.interfaces.FileQueryRepository;
 import deepdivers.community.domain.file.repository.entity.FileType;
 import deepdivers.community.domain.hashtag.controller.interfaces.HashtagQueryRepository;
-import deepdivers.community.domain.like.entity.LikeTarget;
 import deepdivers.community.domain.post.controller.interfaces.ProjectPostQueryRepository;
 import deepdivers.community.domain.post.dto.request.GetPostsRequest;
 import deepdivers.community.domain.post.dto.response.PostPreviewResponse;
@@ -36,7 +34,7 @@ public class ProjectPostQueryRepositoryImpl implements ProjectPostQueryRepositor
     private final FileQueryRepository fileQueryRepository;
 
     @Override
-    public List<ProjectPostPreviewResponse> findAllPosts(Long memberId, GetPostsRequest dto) {
+    public List<ProjectPostPreviewResponse> findAllPosts(final Long memberId, final GetPostsRequest dto) {
         final List<ProjectPostPreviewResponse> postPreviewResponses = extractPostPreview(memberId, dto);
         final List<Long> postIds = postPreviewResponses.stream().map(PostPreviewResponse::getPostId).toList();
         final Map<Long, List<String>> hashtagsByPosts = hashtagQueryRepository.findAllHashtagByPosts(postIds);
@@ -51,12 +49,12 @@ public class ProjectPostQueryRepositoryImpl implements ProjectPostQueryRepositor
         return postPreviewResponses;
     }
 
-    public ProjectPostDetailResponse readPostByPostId(Long postId, Long viewerId) {
+    public ProjectPostDetailResponse readPostByPostId(final Long postId, final Long viewerId) {
         final ProjectPostDetailResponse postDetailResponse = queryFactory
             .select(PostQBeanGenerator.createPostDetail(ProjectPostDetailResponse.class, viewerId))
             .from(post)
             .join(member).on(post.member.id.eq(member.id))
-            .leftJoin(like).on(hasLike(viewerId))
+            .leftJoin(like).on(PostQueryUtils.hasLike(viewerId))
             .where(post.id.eq(postId))
             .fetchOne();
 
@@ -83,13 +81,6 @@ public class ProjectPostQueryRepositoryImpl implements ProjectPostQueryRepositor
             .orderBy(PostQueryUtils.determinePostSortCondition(dto.postSortType()))
             .limit(PostQueryUtils.getLimitOrDefault(dto.limit()))
             .fetch();
-    }
-
-    private BooleanExpression hasLike(final Long memberId) {
-        return post.id
-            .eq(like.id.targetId)
-            .and(like.id.targetType.eq(LikeTarget.POST))
-            .and(like.id.memberId.eq(memberId));
     }
 
 }
