+++ src/test/java/deepdivers/community/domain/post/repository/impl/PostQueryRepositoryImplTest.java
@@ -217,4 +217,28 @@ class PostQueryRepositoryImplTest extends RepositoryTest {
         assertThat(result).hasSizeLessThan(31);
     }
 
+    @Test
+    void 최소_최대_조회개수_사이만큼_조회가_된다() {
+        // given
+        GetPostsRequest dto = new GetPostsRequest(null, null, PostSortType.LATEST, 7);
+
+        // when
+        List<PostPreviewResponse> result = postQueryRepository.findAllPosts(null, dto);
+
+        // then --> deleted data 제외된 결과 값
+        assertThat(result).hasSize(6);
+    }
+
+    @Test
+    void 최소_limit_제한이_있다() {
+        // given, test.sql
+        GetPostsRequest dto = new GetPostsRequest(null, null, null, 2);
+
+        // when
+        List<PostPreviewResponse> result = postQueryRepository.findAllPosts(null, dto);
+
+        // then
+        assertThat(result).hasSize(5);
+    }
+
 }
