+++ src/test/java/deepdivers/community/utils/PostDtoGenerator.java
@@ -2,6 +2,8 @@ package deepdivers.community.utils;
 
 import deepdivers.community.domain.post.dto.response.AuthorInformationResponse;
 import deepdivers.community.domain.post.dto.response.PostPreviewResponse;
+import deepdivers.community.domain.post.dto.response.ProjectPostDetailResponse;
+import deepdivers.community.domain.post.dto.response.ProjectPostPreviewResponse;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -12,6 +14,14 @@ public class PostDtoGenerator {
             1L, 1L, "test", "test", "test", List.of("test"), 0, 0, 0, LocalDateTime.now(),
             new AuthorInformationResponse(1L, "test", "test", "test")
         );
+    }
+
+    public static ProjectPostPreviewResponse generateProjectPostPreview() {
+        return new ProjectPostPreviewResponse(List.of(""));
+    }
+
+    public static ProjectPostDetailResponse generateProjectDetail() {
+        return new ProjectPostDetailResponse();
     }
 
 }
